Renamed "Events" folder to Messages to reflect new hierarchy.
