+++ src/Networking/Messages/SendPassengersCommand.java
@@ -1,4 +1,4 @@
-package Networking.Events;
+package Networking.Messages;
 
 import Networking.Transmitters.DMA_Transmitter;
