+++ test/unit/Networking/DMA_ReceiverTest.java
@@ -1,8 +1,8 @@
 package unit.Networking;
 
 import Networking.Direction;
-import Networking.Events.DestinationEvent;
-import Networking.Events.SystemEvent;
+import Networking.Messages.DestinationEvent;
+import Networking.Messages.SystemEvent;
 import Networking.Receivers.DMA_Receiver;
 import org.junit.jupiter.api.Test;
